Added automatic downloading of benchmark from zenodo and integration into regression test. Also polished regression test to properly report status.

+++ test/run_regression_tests.sh
@@ -4,24 +4,43 @@ set -e
 
 exe=../j-trove.x
 nproc=1
+benchmarks_file=benchmarks.tar.gz
 
-# TODO check exe is present
+# Check exe is present
+if [ ! -f $exe ]; then
+  echo "ERROR: $exe not found."
+  exit 2
+fi
 
-# TODO quiet input or put it somewhere
+# If benchmarks haven't been downloaded, download them
+if [ ! -f $benchmarks_file ]; then
+  echo "Downloading benchmarks"
+  ./download_benchmarks.sh
+  tar xzf $benchmarks_file
+fi
 
-# TODO upload to zenodo
-#./download_benchmarks.sh
-# unzip benchmarks
+# Run series of benchmarks
 for benchmark in H2CO; do
+  # Create output folder & move necessary files
   wd=outputs/$benchmark
   mkdir -p $wd
   cp $exe $wd
   cp benchmarks/$benchmark/input/*.inp $wd
   cp scripts/$benchmark/run_benchmark.sh $wd
+
+  # Run benchmark
   pushd $wd
   echo "Running $benchmark"
   ./run_benchmark.sh $nproc
   popd
+
+  # Compare results with "truth"
   bash scripts/$benchmark/compare_results.sh $wd benchmarks/$benchmark/outputs
-  # TODO actually process the result ( and handle set -e properly in case it fails)
+  success=$?
+  if [ $success -eq 0 ]; then
+    echo "PASS: $benchmark"
+  else
+    echo "FAIL: $benchmark failed with exit code $success"
+    exit $success
+  fi
 done
